Move component to new folder
